+++ app/views/businessactivities/who_is_your_accountant.scala.html
@@ -31,7 +31,7 @@
     heading = header
 ) {
 
-    @form(f, controllers.businessactivities.routes.WhoIsYourAccountantController.post(edit)) {
+    @form(f, controllers.businessactivities.routes.WhoIsYourAccountantNameController.post(edit)) {
 
         @fieldset(
             heading = "businessactivities.whoisyouraccountant.title",
